requirements.txt
fastapi>=0.104.0
requests>=2.31.0
openai>=1.3.0
gspread>=5.12.0
google-cloud-secret-manager>=2.16.4
google-auth>=2.23.0
python-dotenv>=1.0.0
jinja2>=3.1.2
uvicorn>=0.24.0
retrying>=1.3.4
pydub>=0.25.1